Move rms_norm_binary kernel into its own normalization subfolder.
Mirror rms_norm / group_norm_grad layout so kernel and SKILL live under
kernels/normalization/rms_norm_binary/, matching the existing test tree.
