+++ benchmark/one-level-arch/test/kernel/normalization/rms_norm_binary/src/rms_norm_binary.cpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 #include "fileop.h"
-#include "normalization/rms_norm_binary_pto.hpp"
+#include "normalization/rms_norm_binary/rms_norm_binary_pto.hpp"
 
 #ifndef DType
 #define DType __half
